requirements.txt
# فایل: requirements.txt – نسخه نهایی برای VoidVision ONE™

# کتابخانه‌های پایه برای اجرای Streamlit و تحلیل
streamlit>=1.25
numpy
pillow
pandas
matplotlib

# تحلیل تصاویر ماهواره‌ای
rasterio
sentinelhub
shapely
folium

# ساخت گزارش PDF حرفه‌ای
reportlab

# پردازش تاریخ و زمان
python-dateutil

# نمایش نقشه و تعاملی بودن
streamlit-folium

# برای کار با فایل‌های .env اگر نیاز باشد
python-dotenv